+++ packages/store/internals/src/index.ts
@@ -4,8 +4,3 @@ export { memoize } from './memoize';
 export { INITIAL_STATE_TOKEN, InitialState } from './initial-state';
 export { PlainObjectOf, PlainObject, StateClass } from './symbols';
 export { NGXS_STATE_CONTEXT_FACTORY, NGXS_STATE_FACTORY } from './internal-tokens';
-export {
-  localInject,
-  ensureLocalInjectorCaptured,
-  ensureInjectorNotifierIsCaptured
-} from './decorator-injector-adapter';