+++ packages/store/src/decorators/select/select.ts
@@ -1,29 +1,24 @@
-import { ɵivyEnabled } from '@angular/core';
-import { ReplaySubject } from 'rxjs';
-import { mergeMap } from 'rxjs/operators';
-import {
-  localInject,
-  ensureLocalInjectorCaptured,
-  ensureInjectorNotifierIsCaptured
-} from '@ngxs/store/internals';
-
-import { Store } from '../../store';
-import { NgxsConfig } from '../../symbols';
-import { createSelectObservable, createSelectorFn, SelectorFn } from './symbols';
+import { createSelectObservable, createSelectorFn, PropertyType } from './symbols';
 
 /**
  * Decorator for selecting a slice of state from the store.
+ * @deprecated
  */
 export function Select<T>(rawSelector?: T, ...paths: string[]): PropertyDecorator {
   return function(target, key): void {
+    console.warn(
+      `
+      The @Select decorator has been deprecated due to the following reasons:
+      1) Lack of type-safety (compare to 'store.select()')
+      2) Not compatible with server-side rendering because it uses a global 'Store' instance, which might change between renders
+      3) Not compatible with module federation
+      Consider replacing it the with store.select.
+      `
+    );
+
     const name: string = key.toString();
     const selectorId = `__${name}__selector`;
-    let selector: SelectorFn | null = null;
-    let injectorNotifier$: ReplaySubject<boolean> | null = null;
-
-    if (ɵivyEnabled) {
-      injectorNotifier$ = ensureInjectorNotifierIsCaptured(target);
-    }
+    const selector = createSelectorFn(name, rawSelector, paths);
 
     Object.defineProperties(target, {
       [selectorId]: {
@@ -34,33 +29,10 @@ export function Select<T>(rawSelector?: T, ...paths: string[]): PropertyDecorato
       [name]: {
         enumerable: true,
         configurable: true,
-        get() {
-          if (this[selectorId]) {
-            return this[selectorId];
-          }
-          // The `localInject` will be tree-shaken away in apps that
-          // still use the View Engine.
-          if (ɵivyEnabled) {
-            this[selectorId] = injectorNotifier$!.pipe(
-              mergeMap(() => {
-                const store = localInject(this, Store);
-                const config = localInject(this, NgxsConfig);
-                selector = selector || createSelectorFn(config, name, rawSelector, paths);
-                return createSelectObservable(selector, store);
-              })
-            );
-          } else {
-            selector = selector || createSelectorFn(null, name, rawSelector, paths);
-            this[selectorId] = createSelectObservable(selector, null);
-          }
-          return this[selectorId];
+        get(): PropertyType<T> {
+          return this[selectorId] || (this[selectorId] = createSelectObservable(selector));
         }
       }
     });
-
-    // Keep this `if` guard here so the below stuff will be tree-shaken away in apps that still use the View Engine.
-    if (ɵivyEnabled) {
-      ensureLocalInjectorCaptured(target);
-    }
   };
 }