+++ packages/store/src/decorators/select/symbols.ts
@@ -1,8 +1,5 @@
-import { Type, ɵivyEnabled } from '@angular/core';
 import { Observable } from 'rxjs';
 
-import { Store } from '../../store';
-import { NgxsConfig } from '../../symbols';
 import { propGetter } from '../../internal/internals';
 import { SelectFactory } from './select-factory';
 import { StateToken } from '../../state-token/state-token';
@@ -11,36 +8,24 @@ import { throwSelectFactoryNotConnectedError } from '../../configs/messages.conf
 
 const DOLLAR_CHAR_CODE = 36;
 
-export function createSelectObservable<T = any>(
-  selector: any,
-  store: Store | null
-): Observable<T> {
-  // We're doing this stuff to tree-shake the `SelectFactory` when the user
-  // is running Ivy since NGXS will select the state from the provided `store` argument.
-  return ɵivyEnabled
-    ? createSelectObservableIvy(selector, store)
-    : createSelectObservableViewEngine(selector);
+export function createSelectObservable<T = any>(selector: any): Observable<T> {
+  if (!SelectFactory.store) {
+    throwSelectFactoryNotConnectedError();
+  }
+  return SelectFactory.store!.select(selector);
 }
 
-export function createSelectorFn(
-  config: NgxsConfig | null,
-  name: string,
-  rawSelector?: any,
-  paths: string[] = []
-): SelectorFn {
-  rawSelector = rawSelector || removeDollarAtTheEnd(name);
+export function createSelectorFn(name: string, rawSelector?: any, paths: string[] = []): any {
+  rawSelector = !rawSelector ? removeDollarAtTheEnd(name) : rawSelector;
 
-  if (typeof rawSelector !== 'string') {
-    return rawSelector;
-  } else {
+  if (typeof rawSelector === 'string') {
     const propsArray: string[] = paths.length
       ? [rawSelector, ...paths]
       : rawSelector.split('.');
-
-    return ɵivyEnabled
-      ? createSelectorFnIvy(propsArray, config)
-      : createSelectorFnViewEngine(propsArray);
+    return propGetter(propsArray, SelectFactory.config!);
   }
+
+  return rawSelector;
 }
 
 /**
@@ -52,37 +37,8 @@ export function removeDollarAtTheEnd(name: string): string {
   return dollarAtTheEnd ? name.slice(0, lastCharIndex) : name;
 }
 
-export type SelectorFn =
-  | ((state: any, ...states: any[]) => any)
-  | string
-  | Type<any>
-  | StateToken<any>;
-
 export type PropertyType<T> = T extends StateToken<any>
   ? Observable<ExtractTokenType<T>>
   : T extends (...args: any[]) => any
   ? Observable<ReturnType<T>>
   : any;
-
-function createSelectObservableIvy<T = any>(
-  selector: any,
-  store: Store | null
-): Observable<T> {
-  return ngDevMode && !store ? throwSelectFactoryNotConnectedError() : store!.select(selector);
-}
-
-function createSelectObservableViewEngine<T = any>(selector: any): Observable<T> {
-  return SelectFactory.store
-    ? SelectFactory.store.select(selector)
-    : throwSelectFactoryNotConnectedError();
-}
-
-function createSelectorFnIvy(propsArray: string[], config: NgxsConfig | null) {
-  return ngDevMode && !config
-    ? throwSelectFactoryNotConnectedError()
-    : propGetter(propsArray, config!);
-}
-
-function createSelectorFnViewEngine(propsArray: string[]) {
-  return propGetter(propsArray, SelectFactory.config!);
-}