+++ packages/store/src/module.ts
@@ -28,6 +28,7 @@ import { LifecycleStateManager } from './internal/lifecycle-state-manager';
 import { InternalDispatchedActionResults, InternalDispatcher } from './internal/dispatcher';
 import { InternalStateOperations } from './internal/state-operations';
 import { Store } from './store';
+import { SelectFactory } from './decorators/select/select-factory';
 import { StateStream } from './internal/state-stream';
 import { PluginManager } from './plugin-manager';
 import { NgxsRootModule } from './modules/ngxs-root.module';
@@ -64,6 +65,7 @@ export class NgxsModule {
         InternalNgxsExecutionStrategy,
         Store,
         StateStream,
+        SelectFactory,
         PluginManager,
         ...states,
         ...NgxsModule.ngxsTokenProviders(states, options)